ungrab mouse in fullscreen, SDL_VIDEO_X11_DGAMOUSE disabled, More bugfixes
git-svn-id: svn://code.dyne.org/veejay/trunk@769 eb8d1916-c9e9-0310-b8de-cf0c9472ead5

+++ veejay-current/veejay/vj-misc.h
@@ -27,7 +27,7 @@ int   available_diskspace(void);
 
 int vj_perform_screenshot2(veejay_t * info, uint8_t ** src);
 
-float	vj_get_relative_time(void);
+unsigned int	vj_get_relative_time(void);
 
 int	vj_perform_take_bg( veejay_t *info, uint8_t **src);
